# backend/channel_plugin/docs/swagger.yaml
swagger: '2.0'
info:
  title: Zuri Chat Channel Plugin Endpoints
  description: Made By Team Coelho
  termsOfService: https://www.google.com/policies/terms/
  contact:
    email: team-coelho@zuri.chat
  license:
    name: BSD License
  version: http://127.0.0.1:8000
host: localhost:8000
schemes:
  - http
basePath: /api
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Basic:
    type: basic
security:
  - Basic: []
paths:
...
  /v1/{org_id}/channels/:
    get:
      operationId: retrieve-channels
      summary: Get all channels in the organization
      description: |-
        ```bash
        curl -X GET "{baseUrl}/v1/{org_id}/channels/" -H  "accept: application/json"
        ```
      parameters: []
      responses:
        '200':
          description: Response
          schema:
            type: array
            items:
              $ref: '#/definitions/ChannelGet'
        '404':
          description: Error Response
          schema:
            $ref: '#/definitions/Error'
      tags:
        - v1
    post:
      operationId: create-channel
      summary: Create a new channel in the organization
      description: |-
        ```bash
        curl -X POST "{baseUrl}/v1/{org_id}/channels/"
        -H  "accept: application/json"
        -H  "Content-Type: application/json"
        -d "{  "name": "channel name",  "owner": "member_id",  "description": "channel description",  "private": false,  "topic": "channel topic"}"
        ```
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/Channel'
      responses:
        '201':
          description: Response
          schema:
            $ref: '#/definitions/ChannelUpdate'
        '404':
          description: Error Response
          schema:
            $ref: '#/definitions/Error'
      tags:
        - v1
    parameters:
      - name: org_id
        in: path
        required: true
        type: string
...
  /v1/{org_id}/channels/{channel_id}/messages/:
    get:
      operationId: retrieve-channel-messages
      summary: Get all the messages sent in a channel.
      description: |-
        ```bash
        curl -X GET "{{baseUrl}}/v1/{{org_id}}/channels/{{channel_id}}/messages/" -H  "accept: application/json"
        ```
      parameters: []
      responses:
        '200':
          description: Response
          schema:
            type: array
            items:
              $ref: '#/definitions/ChannelMessageUpdate'
        '404':
          description: Error Response
          schema:
            $ref: '#/definitions/Error'
      tags:
        - v1
    post:
      operationId: create-channel-message
      summary: Create a channel message
      description: |-
        ```bash
        curl -X POST "{{baseUrl}}/v1/{{org_id}}/channels/{{channel_id}}/messages/" -H  "accept: application/json"
        ```
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/ChannelMessage'
      responses:
        '201':
          description: Response
          schema:
            $ref: '#/definitions/ChannelMessageUpdate'
        '404':
          description: Error Response
          schema:
            $ref: '#/definitions/Error'
      tags:
        - v1
    parameters:
      - name: channel_id
        in: path
        required: true
        type: string
      - name: org_id
        in: path
        required: true
        type: string
...
  /v1/{org_id}/messages/{msg_id}/:
    get:
      operationId: retrieve-message-details
      summary: Retrieve message details
      description: |-
        ```bash
        curl -X GET "{{baseUrl}}/v1/{{org_id}}/messages/{{msg_id}}/" -H  "accept: application/json"
        ```
      parameters: []
      responses:
        '200':
          description: Response
          schema:
            $ref: '#/definitions/ChannelMessageUpdate'
        '404':
          description: Error Response
          schema:
            $ref: '#/definitions/Error'
      tags:
        - v1
    put:
      operationId: update-message-details
      summary: Update message details
      description: |-
        ```bash
        curl -X PUT "{{baseUrl}}/v1/{{org_id}}/messages/{{msg_id}}/?user_id={{user_id}}&channel_id={{channel_id}}"
        -H  "accept: application/json"
        -H  "Content-Type: application/json"
        -d "{
                "pinned": true, 
                "content": "string"
            }"
        ```
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/ChannelMessageUpdate'
        - name: user_id
          in: query
          description: User ID (owner of message)
          required: true
          type: string
        - name: channel_id
          in: query
          description: Channel ID (ID of channel message was posted)
          required: true
          type: string
      responses:
        '200':
          description: Response
          schema:
            $ref: '#/definitions/ChannelMessageUpdate'
        '404':
          description: Error Response
          schema:
            $ref: '#/definitions/Error'
      tags:
        - v1
    delete:
      operationId: delete-message
      summary: Delete a message
      description: |-
        ```bash
        curl -X DELETE "{{baseUrl}}/v1/{{org_id}}/messages/{{msg_id}}/?user_id={{user_id}}&channel_id={{channel_id}}" -H  "accept: application/json""
        ```
      parameters:
        - name: user_id
          in: query
          description: User ID (owner of message)
          required: true
          type: string
        - name: channel_id
          in: query
          description: Channel ID (ID of channel message was posted)
          required: true
          type: string
      responses:
        '204':
          description: Message deleted successfully
        '404':
          description: Not found
      tags:
        - v1
    parameters:
      - name: msg_id
        in: path
        required: true
        type: string
      - name: org_id
        in: path
        required: true
        type: string
  /v1/{org_id}/messages/{msg_id}/reactions/:
    get:
      operationId: retrieve-message-reactions
      summary: Retrieve message reactions
      description: |-
        ```bash
        curl -X GET "{{baseUrl}}/v1/{{org_id}}/messages/{{msg_id}}/reactions/" -H  "accept: application/json"
        ```
      parameters: []
      responses:
        '200':
          description: Successful
          schema:
            type: array
            items:
              $ref: '#/definitions/ChannelMessageReaction'
      tags:
        - v1
    put:
      operationId: update-message-reactions
      summary: Update message reactions
      description: |-
        ```bash
        curl -X PUT "{{baseUrl}}/v1/{{org_id}}/messages/{{msg_id}}/reactions/"
        -H  "accept: application/json" 
        -H  "Content-Type: application/json"
        -d "{  "title": "string",  "member_id": "string"}"
        ```
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/ChannelMessageReactionsUpdate'
      responses:
        '200':
          description: Reaction updated
          schema:
            type: array
            items:
              $ref: '#/definitions/ChannelMessageReaction'
      tags:
        - v1
    parameters:
      - name: msg_id
        in: path
        required: true
        type: string
      - name: org_id
        in: path
        required: true
        type: string
...
definitions:
  User:
    required:
      - _id
    type: object
    properties:
      _id:
        title: id
        description: User ID
        type: string
        maxLength: 30
        minLength: 1
      role_id:
        title: Role id
        description: Role ID
        type: string
        maxLength: 30
        minLength: 1
      is_admin:
        title: Is admin
        description: 'Default: false. True if the member is an admin'
        type: boolean
        default: false
      notifications:
        title: Notifications
        description: User's notification preferences
        type: object
        additionalProperties:
          type: string
          x-nullable: true
  ChannelGet:
    type: object
    properties:
      _id:
        title: id
        description: Channel ID
        type: string
        readOnly: true
      name:
        title: Name
        description: Channel name
        type: string
        maxLength: 100
        minLength: 1
      description:
        title: Description
        description: Channel description
        type: string
        minLength: 1
      private:
        title: Private
        description: 'Default: false. True if this channel has been set to private.'
        type: boolean
      owner:
        title: Owner
        description: Owner (member_id) of the channel
        type: string
        minLength: 1
      archived:
        title: Archived
        description: 'Default: false. True if this channel has been archived.'
        type: boolean
      topic:
        title: Topic
        description: Channel topic
        type: string
        maxLength: 100
        minLength: 1
      users:
        title: Users
        description: List of users in the channel
        type: object
        additionalProperties:
          type: array
          items:
            $ref: '#/definitions/User'
  BaseError:
    required:
      - message
      - status
    type: object
    properties:
      message:
        title: Message
        type: string
        minLength: 1
      status:
        title: Status
        type: integer
  Error:
    required:
      - error
    type: object
    properties:
      error:
        $ref: '#/definitions/BaseError'
  Channel:
    required:
      - name
      - owner
    type: object
    properties:
      name:
        title: Name
        description: Channel name
        type: string
        maxLength: 100
        minLength: 1
      owner:
        title: Owner
        description: Owner (member_id) of the channel
        type: string
        maxLength: 30
        minLength: 1
      description:
        title: Description
        description: Channel description
        type: string
        minLength: 1
      private:
        title: Private
        description: 'Default: false. True if this channel is set to private.'
        type: boolean
        default: false
      topic:
        title: Topic
        description: Channel topic
        type: string
        maxLength: 100
        minLength: 1
  ChannelUpdate:
    type: object
    properties:
      _id:
        title: id
        description: Channel ID
        type: string
        readOnly: true
      name:
        title: Name
        description: Channel name
        type: string
        maxLength: 100
        minLength: 1
      description:
        title: Description
        description: Channel description
        type: string
        minLength: 1
      private:
        title: Private
        description: 'Default: false. True if this channel has been set to private.'
        type: boolean
      archived:
        title: Archived
        description: 'Default: false. True if this channel has been archived.'
        type: boolean
      topic:
        title: Topic
        description: Channel topic
        type: string
        maxLength: 100
        minLength: 1
...
  ChannelMessageUpdate:
    type: object
    properties:
      _id:
        title: id
        type: string
        readOnly: true
      user_id:
        title: User id
        type: string
        readOnly: true
        minLength: 1
      channel_id:
        title: Channel id
        description: Channel UUID
        type: string
        readOnly: true
        minLength: 1
      can_reply:
        title: Can reply
        type: boolean
        readOnly: true
      type:
        title: Type
        type: string
        enum:
          - message
          - event
        readOnly: true
      edited:
        title: Edited
        description: 'Default: false. True if this message has been updated'
        type: boolean
        readOnly: true
      files:
        description: List of URLs to files/media in this message
        type: array
        items:
          type: string
          x-nullable: true
        readOnly: true
      timestamp:
        title: Timestamp
        type: string
        format: date-time
        readOnly: true
      replies:
        title: Replies
        description: Number of messages sent as replies to this message (threads)
        type: integer
        readOnly: true
      has_files:
        title: Has files
        description: 'Default: false. True if files/media are in this message'
        type: boolean
        readOnly: true
      pinned:
        title: Pinned
        description: 'Default: false. True if this message has been pinned to the
          channel'
        type: boolean
      content:
        title: Content
        description: Body (text) of this message
        type: string
        minLength: 1
      emojis:
        description: List of reactions made to this message
        type: array
        items:
          type: string
          x-nullable: true
        readOnly: true
      event:
        title: Event
        description: Contains the payload, if the 'type' of this object is 'event'
        type: object
        additionalProperties:
          type: string
          x-nullable: true
        readOnly: true
  Event:
    required:
      - action
    type: object
    properties:
      action:
        title: Action
        type: string
        maxLength: 20
        minLength: 1
      recipients:
        type: array
        items:
          type: array
          items:
            $ref: '#/definitions/User'
  ChannelMessage:
    required:
      - user_id
    type: object
    properties:
      user_id:
        title: User id
        type: string
        maxLength: 30
        minLength: 1
      content:
        title: Content
        description: Body (text) of this message
        type: string
        minLength: 1
      files:
        description: List of URLs to files/media in this message
        type: array
        items:
          type: string
          format: uri
          minLength: 1
      event:
        title: Event
        description: Event payload related to this message
        type: object
        additionalProperties:
          $ref: '#/definitions/Event'
      timestamp:
        title: Timestamp
        type: string
        format: date-time
        readOnly: true
...
  ChannelMessageReaction:
    type: object
    properties:
      title:
        title: Title
        description: Emoji title
        type: string
        readOnly: true
        minLength: 1
      count:
        title: Count
        description: Number of reactions made with this emoji
        type: integer
        readOnly: true
      users:
        description: List of users that reacted with this emoji
        type: array
        items:
          type: string
          x-nullable: true
        readOnly: true
  ChannelMessageReactionsUpdate:
    required:
      - title
      - member_id
    type: object
    properties:
      title:
        title: Title
        description: Emoji title
        type: string
        minLength: 1
      member_id:
        title: Member id
        description: User ID
        type: string
        minLength: 1